14655の変更取り込み
